--- baguni-batch/src/main/java/baguni/batch/pojo/resource/WebImage.java
@@ -1,14 +1,0 @@
-package baguni.batch.pojo.resource;
-
-public class WebImage {
-
-	private final WebLink webLink;
-
-	// (+) Image Description
-
-	// (+) Image Alternative
-
-	public WebImage(WebLink webLink) {
-		this.webLink = webLink;
-	}
-}
