+++ baguni-domain/src/main/java/baguni/domain/base/WebImage.java
@@ -1,0 +1,16 @@
+package baguni.domain.base;
+
+import java.net.URL;
+
+public class WebImage {
+
+	private final URL url;
+
+	// (+) Image Description
+
+	// (+) Image Alternative
+
+	public WebImage(URL url) {
+		this.url = url;
+	}
+}
