+++ baguni-domain/src/main/java/baguni/domain/blog/Blog.java
@@ -1,0 +1,15 @@
+package baguni.domain.blog;
+
+import java.net.URL;
+
+import lombok.Getter;
+
+@Getter
+public class Blog {
+
+	private final URL url;
+
+	public Blog(URL url) {
+		this.url = url;
+	}
+}
