+++ baguni-infra/build.gradle
@@ -11,6 +11,7 @@ repositories {
 
 dependencies {
     implementation project(":baguni-common")
+    implementation project(":baguni-domain")
 
     // for @Entity + @Repository
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
