+++ settings.gradle
@@ -1,7 +1,11 @@
 rootProject.name = 'backend'
+
 include 'baguni-common'
-include 'baguni-infra'
 
 include 'baguni-api'
 include 'baguni-batch'
+// ----------------------
+include 'baguni-domain'
+// ----------------------
+include 'baguni-infra'
